requirements.txt
opencv-python
paho-mqtt
numpy
pygame
